working user

+++ exercise-2/app/models/User.scala
@@ -2,7 +2,7 @@ package models
 
 import play.api.libs.json.Json
 
-case class User(var id: Float, var email: String, var password: String)
+case class User(var id: Long, var email: String, var password: String)
 
 object User{
   implicit val userFormat = Json.format[User]
